requirements.txt
anyio==3.7.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
exceptiongroup==1.1.2
Flask==2.3.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pkg_resources==0.0.0
pycparser==2.21
pyOpenSSL==23.2.0
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
python-telegram-bot==20.4
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.4
Werkzeug==2.3.6
zipp==3.16.2
pydantic
